Add gtk composite template support
Known issue: designer signal connection is not working.
