+++ Source/Libs/GtkSharp/Bin.cs
@@ -24,6 +24,9 @@ namespace Gtk {
 	using System.Runtime.InteropServices;
 
 	public partial class Bin {
+
+		protected Bin() : base() { }
+
 		[UnmanagedFunctionPointer(CallingConvention.Cdecl)]
 		delegate IntPtr d_gtk_bin_get_child(IntPtr raw);
 		static d_gtk_bin_get_child gtk_bin_get_child = FuncLoader.LoadFunction<d_gtk_bin_get_child>(FuncLoader.GetProcAddress(GLibrary.Load(Library.Gtk), "gtk_bin_get_child"));
