+++ Source/Libs/GtkSharp/GtkSharp.metadata
@@ -255,6 +255,7 @@
   <attr path="/api/namespace/object[@cname='GtkApplication']/method[@name='GetAccelsForAction']/return-type" name="null_term_array">1</attr>
   <attr path="/api/namespace/object[@cname='GtkApplication']/method[@name='GetActionsForAccel']/return-type" name="null_term_array">1</attr>
   <attr path="/api/namespace/object[@cname='GtkArrow']/method[@name='Set']" name="hidden">1</attr>
+  <attr path="/api/namespace/object[@cname='GtkBin']" name="disable_void_ctor">true</attr>
   <attr path="/api/namespace/object[@cname='GtkBin']/method[@name='GetChild']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkBuilder']/constructor[@cname='gtk_builder_new_from_file']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkBuilder']/constructor[@cname='gtk_builder_new_from_resource']" name="hidden">1</attr>
@@ -331,6 +332,7 @@
   <attr path="/api/namespace/object[@cname='GtkComboBox']/method[@name='SetTitle']" name="name">SetTearoffTitle</attr> 
   <attr path="/api/namespace/object[@cname='GtkComboBox']/signal[@name='Popdown']" name="name">PoppedDown</attr> 
   <attr path="/api/namespace/object[@cname='GtkComboBox']/signal[@name='Popup']" name="name">PoppedUp</attr> 
+  <attr path="/api/namespace/object[@cname='GtkContainer']" name="disable_void_ctor">true</attr>
   <attr path="/api/namespace/object[@cname='GtkContainer']/method[@name='AddWithProperties']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkContainer']/method[@name='ChildGet']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkContainer']/method[@name='ChildGetProperty']" name="hidden">1</attr>
@@ -846,7 +848,9 @@
   <attr path="/api/namespace/object[@cname='GtkViewport']/method[@name='SetVadjustment']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkVScale']/constructor[@cname='gtk_vscale_new_with_range']" name="hidden">1</attr>
   <add-node path="/api/namespace/object[@cname='GtkWidget']"><custom-attribute>[GLib.TypeInitializer (typeof (Gtk.Widget), "ClassInit")]</custom-attribute></add-node>
+  <attr path="/api/namespace/object[@cname='GtkWidget']" name="disable_void_ctor">true</attr>
   <attr path="/api/namespace/object[@cname='GtkWidget']/constructor[@cname='gtk_widget_new']" name="hidden">1</attr>
+  <attr path="/api/namespace/object[@cname='GtkWidget']/method[@name='InitTemplate']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkWidget']/method[@name='ClassFindStyleProperty']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkWidget']/method[@name='ClassListStyleProperties']" name="hidden">1</attr>
   <attr path="/api/namespace/object[@cname='GtkWidget']/method[@name='ClassPath']/*/*[@type='gchar**']" name="pass_as">out</attr>
