+++ Source/Samples/Samples.csproj
@@ -15,7 +15,7 @@
     </EmbeddedResource>
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Sections\**\*.cs" Visible="false">
+    <EmbeddedResource Include="Sections\**\*.cs">
       <LogicalName>GtkSharp.Samples.%(Filename).cs</LogicalName>
     </EmbeddedResource>
   </ItemGroup>
