+++ Source/Templates/GtkSharp.Template.CSharp/content/GtkSharp.CompositeWidget.CSharp/.template.config/template.json
@@ -1,0 +1,33 @@
+{
+  "$schema": "http://json.schemastore.org/template",
+  "author": "GtkSharp Contributors",
+  "classifications": [
+    "Gtk",
+    "UI"
+  ],
+  "name": "Gtk Composite Widget",
+  "identity": "GtkSharp.CompositeWidget.CSharp",
+  "groupIdentity": "GtkSharp.CompositeWidget",
+  "shortName": "gtkcompwidget",
+  "tags": {
+    "language": "C#",
+    "type": "item"
+  },
+  "sourceName": "GtkCompositeWidget",
+  "primaryOutputs": [
+    {
+      "path": "GtkCompositeWidget.cs"
+    },
+    {
+      "path": "GtkCompositeWidget.glade"
+    }
+  ],
+  "defaultName": "GtkCompositeWidget",
+  "symbols": {
+    "namespace": {
+      "description": "Namespace for the generated files",
+      "replaces": "GtkNamespace",
+      "type": "parameter"
+    }
+  }
+}
