+++ Source/Templates/GtkSharp.Template.CSharp/content/GtkSharp.CompositeWidget.CSharp/GtkCompositeWidget.glade
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<interface>
+  <requires lib="gtk+" version="3.18"/>
+  <template class="GtkCompositeWidget" parent="GtkBin">
+    <property name="visible">True</property>
+    <property name="can-focus">False</property>
+    <child>
+      <object class="GtkButton" id="button">
+          <property name="visible">True</property>
+          <property name="can-focus">True</property>
+          <property name="receives-default">True</property>
+        </object>
+    </child>
+  </template>
+</interface>
